REFACTOR: 검증 클래스 생성
검증 과정은 MVC 패턴에서 컨트롤러 영역이라 판단했기 때문에 별도의 클래스를
만들어서 메서드를 모았습니다.
