+++ docs/README.md
@@ -29,26 +29,23 @@
         - 종료 문구 출력
         - 재시작 여부 확인 문구 출력
 
-## 입력 기능
 
-- [x] askNumber
-    - 숫자를 입력받는 메서드
-    - 파라미터: 자릿수
-    - 반환: 입력 문자열
-    - 로직
-        - readLine()으로 입력값 저장
-        - 문자열의 길이가 자릿수와 일치하지 않을 경우 오류 발생
-
-
-- [x] askOneOrTwo
-    - 1 또는 2를 입력받는 메서드
+- askNumber
+    - 숫자 입력 문구를 출력하는 메서드
     - 파라미터: 없음
-    - 반환: 1일 경우 true, 2일 경우 true
+    - 반환: 없음
     - 로직
-        - readLine()으로 입력값 저장
-        - 1 자리 문자열이 아닐 경우 오류 발생
-        - 문자가 1 또는 2가 아닐 경우 오류 발생
+        - 숫자 입력 문구 출력
 
+
+- askOneOrTwo
+    - 1 또는 2 입력 문구를 출력하는 메서드
+    - 파라미터: 없음
+    - 반환: 없음
+    - 로직
+        - 1 또는 2 입력 문구 출력
+
+## 검증 기능
 
 - [x] validateInputLength
     - 입력값의 길이가 유효한지 확인하는 메서드
@@ -64,6 +61,22 @@
     - 반환: 없음
     - 로직
         - 숫자가 1이나 2가 아닐 경우 오류 발생
+
+
+- [x] validateNumber
+    - 숫자가 유효 범위에 속하는지 확인하는 메서드
+    - 파라미터: 숫자 값
+    - 반환: 없음
+    - 로직
+        - 최소 숫자와 최대 숫자 내의 값이 아닐 경우 오류 발생
+
+
+- [x] validateDuplication
+    - 숫자가 중복되는지 확인하는 메서드
+    - 파라미터: 숫자 리스트, 추가하려는 숫자
+    - 반환: 없음
+    - 로직
+        - 숫자 리스트에 추가하려는 숫자가 있을 경우 오류 발생
 
 ## 리스트 생성 기능
 
@@ -87,23 +100,7 @@
         - 정수값이 중복되는지 확인
         - 숫자 리스트에 추가
 
-
-- [x] validateNumber
-    - 숫자가 유효 범위에 속하는지 확인하는 메서드
-    - 파라미터: 숫자 값
-    - 반환: 없음
-    - 로직
-        - 최소 숫자와 최대 숫자 내의 값이 아닐 경우 오류 발생
-
-
-- [x] validateDuplication
-    - 숫자가 중복되는지 확인하는 메서드
-    - 파라미터: 숫자 리스트, 추가하려는 숫자
-    - 반환: 없음
-    - 로직
-        - 숫자 리스트에 추가하려는 숫자가 있을 경우 오류 발생
-
-## 계산 기능
+## 게임 기능
 
 - [x] isNotOver
     - 게임이 종료되는지 확인하는 메서드
